+++ package.json
@@ -14,7 +14,7 @@
     "@ariakit/react": "^0.4.15",
     "@auth/core": "^0.37.4",
     "@auth/dynamodb-adapter": "^2.7.4",
-    "@aws-sdk/client-dynamodb": "^3.712.0",
+    "@aws-sdk/client-dynamodb": "^3.749.0",
     "@aws-sdk/client-s3": "^3.732.0",
     "@aws-sdk/credential-providers": "^3.712.0",
     "@aws-sdk/lib-dynamodb": "^3.321.1",
@@ -84,6 +84,7 @@
     "@udecode/plate-toggle": "^42.0.0",
     "@udecode/plate-trailing-block": "^42.0.0",
     "@vercel/speed-insights": "^1.1.0",
+    "async-mutex": "^0.5.0",
     "axios": "^1.7.9",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
