+++ src/middleware.ts
@@ -4,11 +4,15 @@ import type { NextRequest } from "next/server";
 import { auth } from "@/shared/lib/auth";
 
 export async function middleware(request: NextRequest) {
-  const redirectToLoginPage = async () => {
-    const signInUrl = new URL("/user/signIn", request.url);
-    signInUrl.searchParams.set("from", request.nextUrl.pathname);
+  const session = await auth();
 
-    return NextResponse.redirect(signInUrl);
+  const redirectToLoginPage = (response?: () => NextResponse | undefined) => {
+    if (!session || !session.user || session.error) {
+      const signInUrl = new URL("/user/signIn", request.url);
+      signInUrl.searchParams.set("from", request.nextUrl.pathname);
+      return NextResponse.redirect(signInUrl);
+    }
+    if (response !== undefined) return response();
   };
 
   // API
@@ -17,7 +21,6 @@ export async function middleware(request: NextRequest) {
     request.method === "POST" &&
     request.nextUrl.pathname.startsWith("/api")
   ) {
-    const session = await auth();
     if (!session?.user && !request.nextUrl.pathname.startsWith("/api/user")) {
       return new NextResponse("로그인 후 다시 시도해주세요.", {
         status: 401,
@@ -29,7 +32,6 @@ export async function middleware(request: NextRequest) {
     request.method === "GET" &&
     request.nextUrl.pathname.startsWith("/api/user")
   ) {
-    const session = await auth();
     if (!session?.user) {
       return new NextResponse("로그인 후 다시 시도해주세요.", {
         status: 401,
@@ -44,35 +46,28 @@ export async function middleware(request: NextRequest) {
   // 페이지
   // 관리자 페이지 제한
   if (request.nextUrl.pathname.startsWith("/manage")) {
-    const session = await auth();
-    if (!session?.user) {
-      return redirectToLoginPage();
-    } else if (session?.user?.role !== "ROLE_SUPERUSER") {
-      return new NextResponse("권한이 없습니다.", {
-        status: 403,
-      });
-    }
+    return redirectToLoginPage(() => {
+      if (session?.user?.role !== "ROLE_SUPERUSER") {
+        return new NextResponse("권한이 없습니다.", {
+          status: 403,
+        });
+      }
+    });
   }
 
   // 글쓰기 페이지 제한
   if (request.nextUrl.pathname.startsWith("/post/edit")) {
-    const session = await auth();
-
-    if (!session?.user) {
-      return redirectToLoginPage();
-    }
+    return redirectToLoginPage();
   }
 
   // 로그인 페이지 제한
   if (request.nextUrl.pathname.startsWith("/user")) {
-    const session = await auth();
-
     if (request.nextUrl.pathname.startsWith("/user/signIn")) {
       if (session?.user) {
         // 이미 로그인 돼있을 시 유저 페이지로 리다이렉트
         return NextResponse.redirect(new URL(`/user/myPage`, request.url));
       }
-    } else if (!session?.user) {
+    } else {
       return redirectToLoginPage();
     }
   }
